requirements.txt
dateparser==1.2.0
fastapi==0.115.0
ffmpeg-python==0.2.0
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-auth==2.34.0
httpx==0.27.0
openai>=1.0.0
openai-whisper==20231117
pyobjc>=10.2; platform_system=="Darwin"
pyttsx3==2.90
pywin32>=306; platform_system=="Windows"
pydantic==2.8.2
python-dotenv==1.0.1
regex==2024.7.24
spacy==3.7.4
uvicorn[standard]==0.30.6
google-cloud-texttospeech==2.16.4
